moved non user functions to utils
