+++ src/dagmc_geometry_slice_plotter/__init__.py
@@ -1,10 +1,20 @@
-from importlib.metadata import version
 
-__version__ = version("dagmc_geometry_slice_plotter")
+try:
+    # this works for python 3.7 and lower
+    from importlib.metadata import version, PackageNotFoundError
+except (ModuleNotFoundError, ImportError):
+    # this works for python 3.8 and higher
+    from importlib_metadata import version, PackageNotFoundError
+try:
+    __version__ = version("dagmc_geometry_slice_plotter")
+except PackageNotFoundError:
+    from setuptools_scm import get_version
+
+    __version__ = get_version(root="..", relative_to=__file__)
 
 __all__ = ["__version__"]
 
-from .utils import plot_slice_of_dagmc_file
-from .utils import plot_slice_of_trimesh_object
-from .utils import plot_slice
-from .utils import plot_axis_slice
+from .core import plot_slice_of_dagmc_file
+from .core import plot_slice_of_trimesh_object
+from .core import plot_slice
+from .core import plot_axis_slice
